requirements.txt
fastapi
uvicorn[standard]
pymongo
langchain
langchain-community
langchain-core
langchain-google-genai
faiss-cpu
python-dotenv